pyproject.toml
[project]
name = "one-page-liv-planning"
version = "0.1.0"
description = "LIV Planning FastAPI backend"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-dotenv",
    "passlib[bcrypt]",
    "itsdangerous",
    "bcrypt<4.0",
    "python-multipart",
    "httpx",
    "pytest",
    "requests",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
]
